filter out rendering artifacts
When a line ends on a tile border it is prolonged beyond the usual tile space
to avoid truncated lines from one tile to the other. This code did not take into
account some cornercases which lead to visible drawing artifacts.

The prolonging is only done if the point is at the end of a line. In this case
the line from the neighboring point is taken and extended by a fixed amount. One
obvious case where this went wrong is when the neighbor is so close that it had
the same coordinates in the vector tile. The distance between both points was
then zero, leading to a division by zero and the new point coordinates were
"Nan", resulting in undefined behavior. But even for distinct, but very close
neighbors this led to artifacts.

The code now detects those close neighbors and simply uses the next point in
the line if that isn't within the range limit, then using the next and so on.
If the whole line is within the range limit the entire line is skipped, it would
only result in a single pixel in the generated PNG at the given zoom level.