do not render labels that contain only whitespace
This is not only faster, it also makes sure no other labels can collide with
these invisible labels, so more labels will be shown at the end.

This also does not render whitespace at the beginning and end of labels anymore,
which in turn reduces the collision size of some labels again.