avoid creating temporary String objects for font styles
The values in the MapCSS styles are exactly those we need to put in the font
line, so avoid allocating small new string objects that will live only for a
few lines, but instead just copy the value from the MapCSS style. While at it
switch the comparison to use === as this is only meaningful if both values are
strings, so no need to try any type conversion if they are not.